Control.Concurrent.TxEvent.AckVar

Description

Acknowledgement Variables.

This library provides synchronous acknowledgement variables, which are similar to Id-style synchronous I-variables.

Synchronous acknowledgement variables

data AckVar

The type of acknowledgement variables.

newAckVar :: IO AckVar

Create a new acknowledgement variable.

setAckVar :: AckVar -> IO ()

Asynchronously enable the acknowledgement variable.

getAckVarEvt :: AckVar -> Evt ()

Synchronously query the acknowledgement variables; this event becomes commitable when the acknowledgement variable is enabled.