Control.Concurrent.TxEvent.IO

Description

TxEvent based text output handles

data EvtHandle

An EvtHandle value is much like a Handle value, except that access to the handle is mediated by the Evt monad, rather than the IO monad.

Lifting Handle

mkEvtHandle :: Handle -> IO EvtHandle

Lift a writeable Handle to an EvtHandle.

TxEvent based text output

hPutCharEvt :: EvtHandle -> Char -> Evt ()

Operation hPutCharEvt hdl ch writes the character ch to the file or channel managed by hdl when synchronized upon. See hPutChar.

hPutStrEvt :: EvtHandle -> String -> Evt ()

Operation hPutStrEvt hdl s writes the string s to the file or channel managed by hdl when synchronized upon. See hPutStr.

hPutStrLnEvt :: EvtHandle -> String -> Evt ()

The same as hPutStrEvt, but adds a newline character. See hPutStrLn

hPrintEvt :: Show a => EvtHandle -> a -> Evt ()

Operation hPrintEvt hdl t writes the string representation of t given by the shows function to the file or channel managed by hdl and appends a newline.